techmahedy-4.x: Rename ServiceProvider to Launche
