+++ runtime/config/app.php
@@ -111,18 +111,18 @@ return [
 
     /*
     |--------------------------------------------------------------------------
-    | Application Service Providers
+    | Application Launchers
     |--------------------------------------------------------------------------
     |
-    | Here we register all of the application's service providers. These providers
-    | bootstrap core application services and register important bindings. Providers
+    | Here we register all of the application's launchers. These launchers
+    | bootstrap core application services and register important bindings. Launchers
     | are loaded in the order they are listed below.
     |
-    | Additional providers may be added for specific user implemented features and for packages.
+    | Additional launchers may be added for specific user implemented features and for packages.
     |
     */
-    "providers" => [
-        App\Providers\AppServiceProvider::class,
+    "launchers" => [
+        App\Launchers\AppLauncher::class,
     ],
 
     /*
